Reworked SASS and CSS Framework
Added comments in the SASS files to make it clearer as to what each
file is doing/defining. Also reworked the framework so that sidebar,
defaults, and page SASS files do not have to be edited. When creating a
new page, copy and paste the “index.sass” file and edit the variables.